Add sqlite database backing
This reconstructs boards on startup. That might be too slow — I haven't
tested yet. It has the nice property of reducing the sources of truth
for board data though, so it would be nice to keep.

If it is too slow, it might be parallelizable at least, which would
speed things up a little.