Remove palettes, Simplify object import paths
Removing palettes means requires that each board define it's own
palette. This will likely result in a lot of redundancy, but it makes
more sense that a board's palette is independent from other board
palettes when editing it as a user. To facilitate this, user interfaces
would be best designed when lying to the user about the underlying
structure, using some sort of copy-on-write system.

Such an interface would reduce cases where a user mistakenly edits a
shared palette while assuming that they were only editing a specific
board's palette. It would also make it hard to properly utilize the
underlying data structure.

By harmonizing the underlying data structure with such an interface
design, the temptation of sacrificing UX for properly using the
underlying structure disappears.

This is the sort of thing I might go back on, but hopefully the
reasoning here stands the test of time.