Support non-standard Range requests
This is a save point because I'm probably going to scrap all of this.
Basically, the spec calls for responses to Range headers to return
content in an entirely unsuited format. This is mainly due to the
*required* boundary parameter which is expected to be a delimiter for
data sections. *It is impossible in general to select an appropriate
boundary for arbitrary binary data.*

This doesn't *need* to be a problem. The only reason I can see for this
boundary is if the unit type is not bytes. In this case, the client may
not be able to count the size of the content, and would need a
delimiter.

Of course, that *could* have been solved with a *length* parameter in
the same place or even a Content-Length header in each section (both
of which would have handled arbitrary binary data perfectly), but
apparently that wasn't what was chosen.