Implement placing
And also create a function for connecting to the database and setting
things up, since sqlite apparently doesn't enable foreign key checks by
default 🙄 (and it needs to be done for every connections).

Anyway, this is still missing the ratelimits and such required for an
actual r/place experience, but all the data should be going to the right
places.

There's still some concerns to address over resizing the board and how
positions should change (currently, they don't). Making the board
smaller currently also makes the server unbootable if any placements
fall outside of the new size. These are problems that need solutions,
but they should probably be well thought-out, so I haven't rushed to fix
them yet.