Implement authentication extension
Lots of hardcoded values here for now, but there's no sensitive info and
I'd rather work out a config solution in another commit.

This is getting a bit harder to test since there exists no client
implementation, but it does seem to work as far as I can tell.

I might adjust the spec (and this code) to indicate that clients should
discover the appropriate endpoints from a base instead using oidc since
this presupposes that the client will choose the correct algorithms and
such for authentication.