Send BoardUpdate on place
The BoardUpdate event is lacking; it has no way of indicating when
the board changes in ways other than placements, such as metadata or
mask updates.

More importantly, this implementation is also lacking — I'm not happy
about cloning the socket event for every connection and it would
probably be best if the event was initiated from the Board struct rather
than the route.