Replace endpoint-facing expects with unwraps
And remove the TODO which said to do the opposite. Maybe one day the
server will be able to gracefully fail if the database goes missing, but
the assertion that the database will be available is pervasive in the
codebase now. It's better to let it go at this point and deal with it
all at once later is how I see it.