Implement pixel stacking
Mostly this is just "improve robustness of cooldown calculation". I'm
not entirely happy with this, but I wanted to save-state this code
before I do anything else, because it's likely I'll scrap it before
moving forward.

This approach has a number of flaws which cannot be accounted for
without adding some object state. The most important factor and the one
that other issues stem from is that the cooldown is assumed to be static
in this case. The intention is to eventually support a dynamic cooldown,
so this will not work.

In fact, the only thing that can work consistently with a dynamic
cooldown is storing the value of that cooldown in some form at placement
time. One important thing to consider here is that the base for a
cooldown can often be important to know separately from a final adjusted
value, so it's likely both will need to be stored.

That said, I'll likely know more by the next commit so stay tuned I
guess…