Make cooldown perfectly consistent
Cooldown is now determined based on the placement and its context. This
is done with quite a bit of database querying, which is less than ideal
but also: no extra database columns 🎉.

The `calculate_cooldowns` function is where most of performance is going
to go to repeated lookups, but this can be cached by placement ID. This
cache would be invalidated only on cooldown formula change, but it would
probably be best to flush old values fairly frequently since only up a
users last MAX_STACKED pixels need to be queried for calculating stack
data.