Notify authorized sockets about cooldown events
This took a lot of messing about with to get to work. Mostly due to me
trying work out how many extra actors I would need to do it. Most of
the issue in that probably came from the answer being 0.

In testing, this seems to cover all cases which are important:
- New connections start getting info
- Old connections get updated info after cooldowns have changed
- Current connections get immediate feedback if the current number of
  pixels has changed.

It also seems to give *slightly* incorrect timestamps in testing —
usually off by a few seconds, though this could just be the UTC clock
I'm using being off. Since all testing was done on the same machine,
this is a bit strange however.