Add websocket-based authentication
I was previously pretty adamant that header-based authentication was the
superior way of authenticating a websocket. That was until I realised
that authentication tokens need to expire and — unlike a regular request
— websocket connections are likely to need to deal with this.

Headers cannot solve this problem, so sending authenticate packets
through the socket are the only viable solution.

The good news is that this solves the previous browser incompatibility
problem. The bad news is that socket connections are now even more
complicated. Everything seems to be working well for now, but I can't
help but feel like actix is definitely hurting more than it is helping
here…