Format code
With a moderately configured formatting config file, because I have
opinions and not all of them are shared by those who decided the
defaults. Perhaps I will eventually be swayed to their side — I have
changed my mind on formatting quite a bit over the years, after all.
(I will die on the hill of tabs vs spaces though)

Anyway, the formatter does a bit of a terrible job in some places, but
there's no rules for configuring how it should act in contrived
situations and my planned resolution for when it gets something wrong is
to asses if maybe it's less that the formatter is bad and more that the
code is bad.

Overall, this is small price to pay for automated, consistent
formatting.