Add warp websockets
I was imagining this being a lot prettier in my head. In reality, it's a
bit of a mess. I remain confident that a better solution is available
with this than was available with actix, but this is certainly not it.
There may be some bugs here related to proper closing and cleanup of
websockets but at least now websocket extensions are handled.