Revert to hand-formatting and refactor some things for readiblity
I've decided that I very much don't like how rustfmt decides to format code a
lot of the time. Yes, consistency is good, but I'm not sure it's worth the
ugliness it causes. Relying on a simple tool to solve a complex problem is not
the answer here I think.