Implement separate extensions for board data types
In accordance with the newer draft of the specification, extra board
buffers are extensions. This means that it makes sense to logically
separate out relevant routes and we shouldn't send this data to clients
not requesting those extensions.

The first part was easy, the latter not so much. There's a lot of code
here and I haven't tested any of it, so there may be some bugs. The good
news is that this implementation should be fast and scalable.