Refactor to consolidate database code
This moves all the database code into the appropriate module. This
vastly simpifies the implementation of Board and the code isolation of
things.

A ton of code moved here, but a simple test didn't show any errors
despite the size of the refactor. I wouldn't put it past some mistakes
making it through, but it's a good initial sign at least.

Oh, previous page tokens are gone now too. The field is still there but
they're a pain so I might just remove them from the spec outright.