Use standard directory setup for sea orm migrations
Okay okay, I yield on this one. Being special seemed nice at the time,
but without the extra nested crate the cli can't be used to manage
migrations or add new ones. This made updating the database difficult
since it required running the server, but the server needed the
generated entities to start properly, which can only be obtained from
the database after migration, creating a catch 22 in updating entities.