Implement roles cosmetically
This is mostly just the ldap plumbing to make roles work with the
database, along with the routes for reading them. Permissions are still
not derived and, like users, roles cannot be edited via the server.

That said: this adds an LDAP schema for the required fields for roles.
The base ones included in the x500 spec are insufficient, so it's come
to this. With the lack of referential integrity and questionable
standards for objects, I'm having some doubts about using LDAP. At the
end of the day however, it does seem to be working quite well in
practice once you wade through the masses of iso documents given the
lack of clear information available elsewhere.

Anyway, derived permissions as well as mutability of users and roles
are to come next.