Add permission check to current users
And switch the routing order for current users, matching the special
route before the one for uids.

This permission check is part of the spec. I thought this was erroneous
at first but upon reflection I think my reasoning was that it makes
sense that some implementations may want to hide the user's own endpoint
from themselves. This implementation doesn't really care, but the spec
says "reject without users.current.get" so that's what this does.