Fix accidentaly indentation with spaces
rustfmt doesn't do anything given it's disabled, so some of these creep
in thanks to autocomplete sometimes using them. I hate it, and I hate
everything about this world that space-based indentation not only exists
but is commonplace.

Anyway, rustfmt still mangles the code horribly and I may never truly
recover from having used it before, so it's staying disabled and I
refuse to make the world a worse place by using space indentation so
this is how it hast to be.