Add root events socket
A relatively small diff that felt like pulling teeth. This needs to be
majorly re-organised at some point as it definitely shouldn't live in
the routes module, but that's where it is for now.

I haven't tested this at all, there may well be bugs with it but having
something in place makes it infinitely easier to work on in future at
least.