Implement board notices
This still lacks polish and has the same missing features as site
notices (no mapping of author to user) but works largely the same.

There is probably a lot of room for code deduplication and cleanup here
but this is functional enough for now. It was a bit more work than
expected but not too much.