Use integer user ids in placement table
This adds a mapping of internal numeric ids to string user ids such that
the placement table (and hopefully cached objects in future) weigh far
less and are more easy to deal with. The downside to this is increased
complexity, though hopefully the better storage, memory, and complexity
elsewhere should make up for this in comming commits.