Add density buffer
This allows the cooldown to be fully computed from caches. Since the
databse fetch was around 500µs and is not only a couple of µs, the entire
cooldown calculation when the correct placements are in cache now only
takes about 20µs.

Thanks to all this, pixels per second has gone from about 10 to closer to
500.