Fix packets not being sent to clients subscribing to more events
Events would only be sent to those who subscribed to a non-strict subset
of the events associated with a given board update packet. For example,
if a client subscribed to color, timestamp, and info updates then they
would not be sent and place events because it was sent only to those
subscribing to exactly color, timestamp, or color and timestamp events.

The solution here is to simply build more combinations to send out —
the entire superset, minus the sets that contain no infomation about the
related data. This is perhaps slightly more expensive to handle, but
results indicate that any loss is marginal at best.