Fix lifetime bounds issue for HashLock
By functionally removing the implementation: at least the generic part
of it. Self-referencing generics are trouble, so for now this makes it
concrete which is fine since it was only used in one place.

I'm still not perfectly happy and there's room for a better solution to
be had later.