Fix incorrect density calculation for cooldown
This only considered the density at the current time, which is incorrect
if the examined pixel is not the most recent one at the position.

Calculating this properly requires a database fetch. This should
probably be cached for speed eventually but for now this should
function.

On a related note, since the cooldown function no longer uses sector
data, the sector locks need no longer be passed. This comes with a bonus
— I believe this solves a potential deadlock where a data fetch would
lock the first sector, then a writer would lock the second sector and
try to lock the first while calculating cooldown. Since the first sector
is held by the reader waiting on the writer, and the writer is now
awaiting the reader, they deadlock.

There are likely sitll other cases of this out-of-order sector locking,
and it really is becoming problematic. It is well overdue a significant
shakeup, time allowing.