Replace placement cache with cooldown cache
Similarly to activity, by keeping details specifically about the
cooldown in memory we can always compute cooldown without database
fetches. While this technically still requires some and needs some
trimming around the edges, the core of this works well and allows
instant cooldown lookup without costly cache misses.

The main downside to this is that there is now extra startup cost to
build the cache since it needs to be guarenteed to be correct for it to
function. This requires replaying all pixels which could have
contributed to the current cooldown of users, which is expected to be
far fewer than the entire canvas, but also non-trivial in the worst
case. I expect at most this will introduce a minute or so of delay to
startup.

The other tricky thing to handle regarding this is undos. When cooldowns
before relied on real pixel data to calculate cooldown, simply removeing
the undone pixel from that data was sufficient and the cooldown will be
calculated correctly. Now that your cooldown relies on a
proof-by-induction, changing past data requires re-running our
calculations. Fortunately, we can do this if we simply hold on to past
cache data as long as it could be undone, then we can simply pop the
top data, remove the entry, and re-insert the data with the correct
calculations.

I have admittedly not tested that, but it is sound in theory.