improve usercount lookup time from O(n) to O(1)
Previously in order to count the number of active users we had to
iterate through all pixels which contribute to this. An alternative
structure which has a constant time to count unique entries is a map. A
map of users to the most recent place time would work, but would have a
downside: there is no trivial way to find the lowest value in the map,
so while computing goes from O(n) to O(1), culling — which is part of
the same step — goes from ~O(1) to O(n).

Instead of doing one or the other, the trick here is to do both: we can
use the sorted list of placements to address the map and
increase/decrease the correct key when appropriate, storing the current
number of user pixels in that buffer as the map value. This map can then
be queried instantly for its size, and the buffer can be culled quickly,
since is sorted.