Implement placement buffering
This sends placements to the database on a seperate thread for a
significant speedup in placement time, comitting to a placement before
it hits the database.

This means that an unexpected loss of power can result in a loss of the
immediate data from the last few milliseconds in the worst case, but
given the unlikeliness and inseverity of this, it seems an acceptable
risk for better database pressure management and a significantly reduced
critital section time (2ms down to just 20µs in testing).