Fix infinite loop when converting BoardsDatabaseError to StatusCode
In debug mode this would likely result in a stack overflow due to
recusion limits. Unforunately, the compiler is smart enough to optimize
this to a busy loop in release mode, which lock the program and is
difficult to track down.

At least, I'm really hoping that was the issue.