Remove LDAP
LDAP was, in theory, an interesting idea. By seperating all of the user
data from the strictly board data, a more microservices architecture
could have been implemented to access this data. The idea behind using
LDAP was that this data would have a standardized form and applications
would generally be compliant with other data stores.

As the project went on, it became more apparent that LDAP does not quite
support enough of a standard to bare this out in reality. Custom schemas
were drafted, more and more objects went into the SQL database as the
need for structured data grew, and the pains of maintaining two seperate
database systems became apparent.

With this, I have decided that it is indeed best to suffer an SQL
solution for all data types, given whatever choice is made that the
schema will be custom. This is very healthy for the codebase I believe —
code reuse and standardization is now a lot more feasible and indeed
already a lot of good refactoring at the database level has been
implemented in this commit. I hope that more will come in the future.

Of course, the removal of an old database requires the migration of the
data to maintain a cohearent experience, and so m12 will connect and
transfer all LDAP data correctly to SQL when the approproate ENV config
variables are set. Since this requires keeping the dependancy on LDAP as
a library, it has been gated behind the "migrate-ldap" feature so that
regular builds can benefit from the decreased code complexity. When the
feature is not enabled, the system will act as though LDAP is empty and
setup the tables as normal.

Hopefully this is a good first step to reducing some of the technical
debt that has been accrued from thius misstep.