Break database into modules by object and introduce specifiers
Specifiers may be a bit verbose, but the intent is to make path parsing
and generation generic and consistent. The implementation may be a
little half-baked here, but we'll see how it pans out. A lot of the
boilerplate for them could probably be removed with a good proc-macro,
but that has it's own issues which I'd rather not tackle in this commit.